refactor(name-version): Moving name version packages from project.
