+++ src/main/java/com/synopsys/integration/detect/tool/detector/DetectorTool.java
@@ -38,8 +38,8 @@ import org.slf4j.LoggerFactory;
 
 import com.synopsys.integration.detect.exception.DetectUserFriendlyException;
 import com.synopsys.integration.detect.exitcode.ExitCodeType;
-import com.synopsys.integration.detect.kotlin.project.DetectorEvaluationNameVersionDecider;
-import com.synopsys.integration.detect.kotlin.project.DetectorNameVersionDecider;
+import com.synopsys.integration.detect.kotlin.nameversion.DetectorEvaluationNameVersionDecider;
+import com.synopsys.integration.detect.kotlin.nameversion.DetectorNameVersionDecider;
 import com.synopsys.integration.detect.lifecycle.shutdown.ExitCodeRequest;
 import com.synopsys.integration.detect.tool.detector.impl.ExtractionEnvironmentProvider;
 import com.synopsys.integration.detect.workflow.event.Event;
