+++ src/main/kotlin/com/synopsys/integration/detect/kotlin/nameversion/DetectorProjectInfo.kt
@@ -1,4 +1,4 @@
-package com.synopsys.integration.detect.kotlin.project
+package com.synopsys.integration.detect.kotlin.nameversion
 
 import com.synopsys.integration.detector.base.DetectorType
 import com.synopsys.integration.util.NameVersion
