+++ src/main/kotlin/com/synopsys/integration/detect/kotlin/nameversion/NameVersionDecisions.kt
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.detect.kotlin.project
+package com.synopsys.integration.detect.kotlin.nameversion
 
 import java.util.Optional
